server/requirements.txt
black==19.3b0
django==2.1.5
psycopg2-binary==2.7.6.1
Pillow==5.3.0
django-cors-headers==2.4.0
django-jwt-auth==0.0.2
factory-boy==2.11.1
PyJWT==1.7.1
gevent==1.4.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
django-extensions==2.1.4
ipdb==0.11
serious-django-graphene==0.1
django-graphql-jwt==0.2
celery==4.2.1
redis==3.2.0
flower==0.9.2
pytest==4.3.0
pytest-cov==2.6.1
pytest-django==3.4.8
snapshottest==0.5.0
tornado>=4.2.0,<6.0.0
